requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
idna==2.10
lxml==4.5.2
pip==20.3.1
PyExecJS==1.5.1
requests==2.24.0
setuptools==49.6.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.10
